+++ py/teardown.py
@@ -1,6 +1,6 @@
 from PyQt5.QtCore import QCoreApplication
 
-from web_server.screen_cast import stopScreenCastServer
+from webserver.screen_cast import stopScreenCastServer
 from audio_playback import stopAudioPlayback
